Fixed virtual switch interface_address and os_managed fields
The interface_address was being setted every puppet run because the
underlying interface can have two address (IPv6 and IPv4) and the user
only specifies one of them.Now the address specified in the manifest
is the one returned if it's bound to the virtual switch defined in the
system.

The os_managed PS script was returning the string 'True' or 'False',
now this value is checked in ruby to return proper boolean values.