Creates nova-compute windows service and defines the depencencies
This patch adds a new defined resource to handle windows services
based on Python. It makes use of registry module to create the
service and additionals entries required for creating a service
running a python script.

The nova-compute service is created using this defined resource,
it also configures the folders and dependencies.